track success

+++ src/track_pkg/scripts/controller/lat_controller.py
@@ -49,7 +49,7 @@ class LatController():
         # lookahead = min(self.k * self.ego.speed + self.lookahead_default, 6)
         lookahead = 6
         # target_index = lookahead*10 + self.ego.index
-        target_index = lookahead*self.ego.speed + self.ego.index
+        target_index = int(lookahead * self.ego.speed) + self.ego.index
         
         target_x, target_y = self.path.x[target_index], self.path.y[target_index]
